main/requirements.txt
parliament-functions==0.1.0
flask
requests
elementpath
boto3